settings: enable configuration of IO buffer size
Settings subsystem was giving the possibility to define the size of the
buffer used for accessing the storage medium. This size was hard-coded
to 16 bytes. This commit enables the user to configure it, which can be
needed when the storage medium requires larger read/write alignment or
can be desirable to increase read/write performance for example.

Signed-off-by: Thomas Altenbach <taltenbach@witekio.com>